refactor : MVC 패턴 리팩토링
