+++ src/main/java/baseball/config/Config.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.config;
 
 public class Config {
